Add python API for pulling tickers with filters
The pulling functionality allows to:
 - Pull any given amount of tickers via the `yfinance` package.
 - Specify filter criteria which work on a tickers historical data
   price data. The filters can be added dynamically and allows any
   number of filters.
 - Define a start data for historical data.
 - Filter the pulled data online to ensure the memory size does not
   explode when pulling data for many ticekrs.

This change implements a single filter which implements the request
short squeeze filter:
`definition of a major short squeeze is when a stock
doubles in price (or more) within one week`

This commit also introduces integration tests and corresponding
pytest markers to dynamically select tests based on a
developer's needs.

**Testing**
The code was solely tested via the newly written unit and
integration tests. As we have no CI in place,
here are the test results.
Unit tests:
```
$ poetry run python -m pytest -m "not integration_test" test/
...
============== 5 passed, 2 deselected in 0.38s ==============
```

Integration tests:
```
$ poetry run python -m pytest -m "integration_test" test/
...
============== 2 passed, 5 deselected in 1.58s ==============
```