Create analysis data to answer q4
This change is not sane. The filter `multiply_price_within_x_days`
function was changed and is now broken when using it on downloaded
data. After storing the data as csv and loading them again the
filter works. For the creation of the data this is anyways the
workflow and due to time constraints I am now pushing this change.
I will clean this up in the next 24 hours.

**Testing**
Unit and integration tests (one test now additionally fails):
```
$ poetry run python -m pytest test/
...
=============== 12 passed, 2 xfailed in 8.39s ===============
```