Make `get_tickers_fixed.py` distinguish ecxchanges
The supposed 'sane' module taken from this issue [1] is broken. It
always returns all tickers even if only a single exchange is selected.

This commit fixes this problem by using the implementation from this
pull request [2].

[1] https://github.com/shilewenuw/get_all_tickers/issues/12
[2] https://github.com/shilewenuw/get_all_tickers/pull/17/commits

**Testing**
Unit and integration tests:
```
$ poetry run python -m pytest test/
...
========== 12 passed, 2 xfailed, 3 warnings in 6.85s ===========
```

Manually pulling the tickers with the API per exchange. The
ticker count return for the different exchanges is now:
 - NYSE: 2671
 - NASDAQ: 4249
 - AMEX: 252
 - All enabled: 7172