Unify historical price panda's data frames
Previously, the serialization and deserialization of ticker
price data lead to a panda layout different to the one
when the data was freshly downloaded.

This commit resolves the mismatch. The downloaded data and
the same data csv-loaded bo succeed in equality checks

This change reenables all tests again.
the way how the historical ticker price data was serialized
and deserialzied

**Testing**
Unit and integration tests:
```
$ poetry run python -m pytest test/
...
=================== 14 passed in 7.01s  ====================
```